Changed Config Management & Restructure
