+++ .gitignore
@@ -8,3 +8,4 @@ config.json
 .Trashes
 ehthumbs.db
 Thumbs.db
+config.env
