+++ config.env.example
@@ -1,0 +1,13 @@
+#Dualis Credentials
+DUALIS_USER=dualis@login.de
+DUALIS_PASSWORD=9876qwertz
+
+#Polling Interval
+INTERVAL=25
+
+#SMTP Credentials
+SMTP_HOST=lehre-mail.dhbw-stuttgart.de
+SMTP_PORT=587
+SMTP_USER=1234@567.de
+SMTP_PASSWORD=myPassword
+SMTP_RECEIVER=myGrades@changeMe.de
