--- config.json.template
@@ -1,13 +1,0 @@
-{
-    "username": "",
-    "password": "",
-
-    "SMTPHost": "",
-    "SMTPPort": 587,
-    "SMTPUsername": "",
-    "SMTPPassword": "",
-
-    "notificationRecipient": "",
-
-    "updateIntervalMinutes": 25
-}
